The last pillar is a "glitched one" and will finish the game when you…

+++ include/helpers.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <stdbool.h>
 
-bool isColliding(float x1, float y1, int w1, int h1, float x2, float y2, int w2, int h2);
+bool isColliding(float x1, float y1, float w1, float h1, float x2, float y2, float w2, float h2);
